pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = ["uicodee <uicodee@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.29.0"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
pydantic = {extras = ["dotenv"], version = "^2.7.1"}
pytz = "^2024.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.1.3"
python-multipart = "^0.0.9"
greenlet = "^3.0.3"
alembic = "^1.13.1"
pydantic-settings = "^2.4.0"
boto3 = "^1.35.5"
aiohttp = "^3.10.5"
sse-starlette = "^2.1.3"
pyjwt = "^2.9.0"
cryptography = "^43.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
